+++ lib/glimmer/swt/combo_proxy.rb
@@ -99,6 +99,8 @@ module Glimmer
           }
         }.to_s
       end
+      
     end
   end
+  
 end